requirements.txt
pimoroni-bme280==0.1.1
smbus2==0.4.0
flask==1.1.2
flup==1.0.3
arrow==0.17.0
